requirements.txt
botasaurus==4.0.88
python-telegram-bot==20.6
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
yandexcloud==0.236.0